Fixed Some Runtime Issues When running the tool with no config specified, the std default derive would not use serde defaults. This has been alleviated by implementing Default manually.
In Schema mode, when files are not produced, like input_objects.rs, the mod.rs still included them in the imports.
As far as I can tell, the code is already prepared to return false when this happens, so I just removed the hardcoded true that was there, as that would override the file creation bool as true regardless of what the function reported.

Finally, I updated all the dependencies in the main project. I didn't get around to updating the dependencies in the examples however.